services/skin-analysis-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
motor>=3.3.2
pymongo>=4.6.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.22.0
aiofiles>=23.2.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
pytest-cov>=4.1.0
pytest-mock>=3.12.0
psutil>=5.9.6
redis>=5.0.0

# ML Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0
huggingface-hub>=0.16.0